[#473] Tag 엔티티 수정 및 Cell과 연결
