+++ Domain/Interfaces/DynamicCell.swift
@@ -1,0 +1,16 @@
+//
+//  DynamicCell.swift
+//  Reazy
+//
+//  Created by 문인범 on 2/19/25.
+//
+
+import Foundation
+
+
+protocol DynamicCell: Hashable, Identifiable {
+    var id: UUID { get }
+    var name: String { get }
+    
+    func getCellWidth() -> CGFloat
+}
