+++ Presentation/Home/HomeSearch/Cell/PDFTagCell.swift
@@ -30,32 +30,3 @@ struct PDFTagCell<Tag: DynamicCell>: View {
         }
     }
 }
-
-#Preview {
-    HStack {
-        PDFTagCell(tag: TemporaryTag.init(name: "test")) {}
-        PDFTagCell(tag: TemporaryTag.init(name: "testfdas")) {}
-        PDFTagCell(tag: TemporaryTag.init(name: "testggggg")) {}
-        PDFTagCell(tag: TemporaryTag.init(name: "test")) {}
-        PDFTagCell(tag: TemporaryTag.init(name: "test")) {}
-    }
-}
-
-// TODO: 추후 엔티티 수정 및 폴더링 예정
-struct TemporaryTag: DynamicCell {
-    let id = UUID()
-    let name: String
-    
-    public func getCellWidth() -> CGFloat {
-        let count = self.name.count
-        return CGFloat(10 + count * 10)
-    }
-}
-
-
-protocol DynamicCell: Hashable, Identifiable {
-    var id: UUID { get }
-    var name: String { get }
-    
-    func getCellWidth() -> CGFloat
-}
